refactor: move events & dimensions code to separate classes
